[refactor] #156 예약 도메인 서비스, 컨트롤러 구조 변경
